# app/static/swagger.yaml
swagger: '2.0'
info:
  title: 'CryptoTemple API'
  description: 'Cryptocurrency trading simulation platform API'
  version: '1.0.0'
host: 'localhost:5000'
schemes:
  - 'http'
consumes:
  - 'application/json'
produces:
  - 'application/json'
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: 'JWT token for authentication (format: Bearer <token>)'
paths:
...
  /settings/{user_id}/profile:
    get:
      tags:
        - Settings
      summary: "Get user profile"
      description: "Retrieve user profile information"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "user_id"
          type: integer
          required: true
          description: "User ID"
      responses:
        200:
          description: "Profile retrieved successfully"
          schema:
            $ref: "#/definitions/SettingsProfileResponse"
        403:
          description: "Unauthorized access"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "User not found"
          schema:
            $ref: "#/definitions/ErrorResponse"

    put:
      tags:
        - Settings
      summary: "Update user profile"
      description: "Update username and/or email (rate limited to 10 per minute)"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "user_id"
          type: integer
          required: true
          description: "User ID"
        - in: "body"
          name: "body"
          description: "Profile updates"
          required: true
          schema:
            $ref: "#/definitions/UpdateProfileRequest"
      responses:
        200:
          description: "Profile updated successfully"
          schema:
            $ref: "#/definitions/UpdateProfileResponse"
        403:
          description: "Unauthorized access"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "User not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
        409:
          description: "Username or email already taken"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /settings/{user_id}/change-password:
    post:
      tags:
        - Settings
      summary: "Change password"
      description: "Update user password (rate limited to 5 per minute)"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "user_id"
          type: integer
          required: true
          description: "User ID"
        - in: "body"
          name: "body"
          description: "Password information"
          required: true
          schema:
            $ref: "#/definitions/ChangePasswordRequest"
      responses:
        200:
          description: "Password changed successfully"
          schema:
            $ref: "#/definitions/MessageResponse"
        401:
          description: "Current password is incorrect"
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: "Unauthorized access"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "User not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /settings/{user_id}/reset-balance:
    post:
      tags:
        - Settings
      summary: "Reset balance"
      description: "Reset balance to $10,000 and clear holdings (rate limited to 3 per day)"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "user_id"
          type: integer
          required: true
          description: "User ID"
        - in: "body"
          name: "body"
          description: "Confirmation"
          required: true
          schema:
            $ref: "#/definitions/ConfirmationRequest"
      responses:
        200:
          description: "Balance reset successfully"
          schema:
            $ref: "#/definitions/ResetBalanceResponse"
        400:
          description: "Confirmation required"
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: "Unauthorized access"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "User not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /settings/{user_id}/delete-account:
    delete:
      tags:
        - Settings
      summary: "Delete account"
      description: "Permanently delete user account (rate limited to 1 per day)"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "user_id"
          type: integer
          required: true
          description: "User ID"
        - in: "body"
          name: "body"
          description: "Account deletion confirmation"
          required: true
          schema:
            $ref: "#/definitions/DeleteAccountRequest"
      responses:
        200:
          description: "Account deleted successfully"
          schema:
            $ref: "#/definitions/MessageResponse"
        400:
          description: "Confirmation required"
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: "Password is incorrect"
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: "Unauthorized access"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "User not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /settings/{user_id}:
    get:
      tags:
        - Settings
      summary: "Get all settings"
      description: "Retrieve all user settings"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "user_id"
          type: integer
          required: true
          description: "User ID"
      responses:
        200:
          description: "Settings retrieved successfully"
          schema:
            $ref: "#/definitions/AllSettingsResponse"
        403:
          description: "Unauthorized access"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "User not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /settings/{user_id}/security:
    get:
      tags:
        - Settings
      summary: "Get security settings"
      description: "Retrieve security-related information"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "user_id"
          type: integer
          required: true
          description: "User ID"
      responses:
        200:
          description: "Security settings retrieved"
          schema:
            $ref: "#/definitions/SecuritySettingsResponse"
        403:
          description: "Unauthorized access"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "User not found"
          schema:
            $ref: "#/definitions/ErrorResponse"

# ============================================================================
# DEFINITIONS
# ============================================================================
definitions:
...
  SettingsProfileResponse:
    type: object
    properties:
      id:
        type: integer
      username:
        type: string
      email:
        type: string
      cash_balance:
        type: number
        format: float
      is_active:
        type: boolean
      created_at:
        type: string
        format: date-time
      last_login:
        type: string
        format: date-time
        nullable: true
  UpdateProfileRequest:
    type: object
    properties:
      username:
        type: string
        example: "new_username"
      email:
        type: string
        format: email
        example: "newemail@example.com"
  UpdateProfileResponse:
    type: object
    properties:
      message:
        type: string
      user:
        type: object
        properties:
          id:
            type: integer
          username:
            type: string
          email:
            type: string
  ChangePasswordRequest:
    type: object
    required:
      - current_password
      - new_password
    properties:
      current_password:
        type: string
        format: password
      new_password:
        type: string
        format: password
        minLength: 8
  ResetBalanceResponse:
    type: object
    properties:
      message:
        type: string
      cash_balance:
        type: number
        format: float
  ConfirmationRequest:
    type: object
    required:
      - confirm
    properties:
      confirm:
        type: boolean
        example: true
  DeleteAccountRequest:
    type: object
    required:
      - password
      - confirm
    properties:
      password:
        type: string
        format: password
      confirm:
        type: boolean
        example: true
  AllSettingsResponse:
    type: object
    properties:
      profile:
        type: object
        properties:
          id:
            type: integer
          username:
            type: string
          email:
            type: string
          created_at:
            type: string
            format: date-time
      account:
        type: object
        properties:
          is_active:
            type: boolean
          last_login:
            type: string
            format: date-time
      trading:
        type: object
        properties:
          cash_balance:
            type: number
            format: float
  SecuritySettingsResponse:
    type: object
    properties:
      account_security:
        type: object
        properties:
          last_login:
            type: string
            format: date-time
          account_active:
            type: boolean
          created_at:
            type: string
            format: date-time
      available_actions:
        type: array
        items:
          type: string
        example: ["change_password", "reset_balance", "delete_account"]

  # Common Definitions
  ErrorResponse:
    type: object
    properties:
      message:
        type: string
        example: "Error message"
  MessageResponse:
    type: object
    properties:
      message:
        type: string
        example: "Success"